pyproject.toml
[project]
name = "edge-sensor-gateway"
version = "0.1.0"
description = "Industrial edge sensor gateway simulator built with Python, asyncio, FastAPI, MQTT and WebSocket."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "PyYAML>=6.0.2",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "paho-mqtt>=2.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]